Bump dates of last two posts back by one day
I actually wrote both of these on 2024-06-03.  But currently
I have only listed the date in markdown without any timestamp.
The easy solution is to put them on separate days.  But apparently
wherever GitHub is building my website, it is not yet June 4.
So the latest post is not showing up.

The next easy solution is to bump the days backwards by one day.
Who will notice anyways =)